keras-yolo/requirements.txt
Baker==1.3
bleach==1.5.0
cytoolz==0.9.0.1
futures==3.1.1
h5py==2.7.1
imgaug==0.2.5
Keras==2.1.1
lxml==4.2.1
matplotlib==2.1.0
numpy==1.13.3
path.py==11.0.1
progressbar2==3.34.3
scikit-learn==0.19.1
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
toolz==0.9.0
